some database upsert debugging

+++ src/Database/postgresql/1-dbCrea-BalanceLogArchiveMain-pgSQL.sql
@@ -1,8 +1,8 @@
 CREATE FOREIGN TABLE balancelog_archive
 (
-  yearlog INT,
-  monthlog INT,
-  daylog INT,
-  hourlog INT,
-  balance FLOAT
+  yearlog SMALLINT,
+  monthlog SMALLINT,
+  daylog SMALLINT,
+  hourlog SMALLINT,
+  balance DECIMAL
 ) SERVER archivesrv OPTIONS ( table_name 'balancelog', use_remote_estimate 'true', updatable 'true' );
