+++ src/Database/postgresql/1-dbCrea-HourlyBalCountMain-pgSQL.sql
@@ -1,11 +1,11 @@
 CREATE TABLE hourlybalcount
 (
   hcreationtime TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-  hyearlog INT,
-  hmonthlog INT,
-  hdaylog INT,
-  hhourlog INT,
-  hbalance FLOAT,
+  hyearlog SMALLINT,
+  hmonthlog SMALLINT,
+  hdaylog SMALLINT,
+  hhourlog SMALLINT,
+  hbalance DECIMAL,
   CONSTRAINT PK_HOURLYBALCOUNT_CREATION PRIMARY KEY (hcreationtime)
 ) WITH (autovacuum_enabled = 'true', autovacuum_vacuum_threshold = 5);
